gce-xfstests ltm: delete dead VMs unless KEEP_DEAD_VM option present.
When a shard process detects that its test appliance is wedged (by a
timeout), the shard dumps its serial data instead of results. Usually if
a test VM is wedged, keeping the VM is a moot point as the kernel can
have crashed entirely.

This change deletes a timed out VM by default, setting a shutdown reason
of "ltm detected test timeout". The shard will still wait for the VM to
be fully deleted, and then look for a results file.  If it does actually
find a results file, this indicates that the VM was in a state that was
still able to execute the shutdown script properly, and these results
will be downloaded in addition to adding in the serial data dump.

Setting GCE_LTM_KEEP_DEAD_VM="<non empty string>" will prevent the
shutdown entirely, and upon detecting timeout, the shard will dump
serial output and stop monitoring the VM, leaving it be.

Signed-off-by: Jason Liu <liujason@google.com>
Google-Bug-Id: 64598993
Change-Id: If2ed64f46c79cad1e5927b6c1609851d12f056b8