gce-xfstests ltm: don't upload results if no shards were even started
In the event that the gce-xfstests command line fails to run in all
shard processes, or if all shard processes end up not creating a serial
dump or a results directory, the testrun should not upload an empty
results directory. Instead, error-level logs will be logged to the
shard, and nothing will be created.

Signed-off-by: Jason Liu <liujason@google.com>
Change-Id: Ia69ab4e15c959f45f2824fc9b791020fac44ead1