gce-xfstests ltm: parse "ltm" as gce-xfstests specific command
The original implementation of ltm used the "--ltm" flag anywhere in the
command line to specify that the entire command line should be sent to
the LTM. This was slightly confusing.

This change moves and renames the "--ltm" command option to "ltm",
putting it in gce-xfstests specifically and not as an option flag in
parse_cli. The rest of the behavior of sending the command to the LTM is
the same.

The command is now required to be the first argument to gce-xfstests,
similarly to "gce-xfstests setup" and "gce-xfstests launch-ltm",
otherwise it will be treated as an argument to util/parse_cli (and fail
the command, as "ltm" is not a regular option, and it is also not a
valid test name).

Signed-off-by: Jason Liu <liujason@google.com>
Change-Id: I98a166a32015b74110909e2cdea1df0c8928b403