gce-xfstests ltm: wait for shorter time for results in shard
When the shard process is looking for the results URI, it polls the
bucket for 5 minutes, with 10 seconds in between each poll (and a 10
second delay before the first poll).

This is likely unnecessary. If the shard process is looking for the
results URI, the test VM has already been fully deleted. If at this
point, the results are not in the bucket, it is unlikely that the
results will ever show up in the bucket after the fact.

This change causes the shard to wait for only 20 seconds, polling a
total of 5 times (once immediately, and then 4 more times delayed by 5
seconds each). If the results tarball can't be found within that time,
the same behavior as before will occur (the serial dump will be used in
place of the results).

Signed-off-by: Jason Liu <liujason@google.com>
Change-Id: Ib4d615823735cdd8caa6600f9689177e9d6f97ce