gce-xfstests ltm: properly deal with --kernel option on ltm
When the user specifies a --kernel, the local command line either
discovers that the kernel option specified is a target in the GS bucket
(which then sets GS_KERNEL equal to it), or that the kernel option is a
locally stored file. It then creates a GS_KERNEL of bzImage-*-onetime,
and uploads the local file to the GS_KERNEL location.

This change allows the local command line to send the GS_KERNEL option
to the LTM, and have the LTM discover it and use it.

As the command line will contain the original --kernel option, the LTM
needs to remove the option from the command line and append its own, if
the gs_kernel is sent over. Otherwise, the --kernel option may fail if
the original user passed a local file into the option, because the LTM
will not have that path present as a file.

Signed-off-by: Jason Liu <liujason@google.com>
Change-Id: I58214656cddbaf82894cb8f395dadb860e606a26