ltm: don't expand 'quick' or 'full' aliases in cmdparser
Expanding aliases in the cmdparser is necessary for 'smoke' as it turns
into '-c 4k -g quick'. It isn't necessary for quick or full, at least
not now, as test VM division is only done on a per-config basis, and
doesn't split the FSTESTSETs apart.

This change removes the unnecessary expansion of the 'quick' and 'full'
aliases, as 'gce-xfstests ltm -g quick' was causing the LTM to execute a
gce-xfstests command with an empty -g option (quick on its own was being
expanded to '-g quick' and prepended to the argument list; the command
in the end looked something like
'gce-xfstests <other opts> -g quick -g').

Signed-off-by: Jason Liu <liujason@google.com>
Change-Id: Ic16b6671c1a57c22f4b063a282f0fbe33943cfe9