gce-xfstests: don't upload kernel if path is gs://*
In uploading the kernel image, an OR check with RUN_ON_LTM was added
earlier to ensure that the GS bucket would contain an image for the LTM
to use.

If the GCE_KERNEL variable is empty however (e.g. user specifies a gs://
path directly to the --kernel option), the script will fail as it won't
be able to find the empty string as a file.

This change fixes that.

Signed-off-by: Jason Liu <liujason@google.com>
Change-Id: Ic02e3f7e9809b2f1bff0782da109bd34f8d7ffa4