ltm: send email of aggregate test run
The LTM originally didn't send an email, and if report emailing was
configured, the individual shards would each send their own email
report (could be a bit spammy)

This change allows the LTM to send the email report if the sendgrid API
key is configured. Additionally, an '--email' command line option has
been added. This command-line option specifies the receiver of the
report. If the next argument passed is an empty string (--email ''), the
testrun will not send a report. This also allows the LTM to explicitly
tell its shards not to send individual emails.

The email report will contain the aggregate failures file that the LTM
provides in its results tarball, which includes the failures file of
each of its shards concatenated end to end with some shard headers in
between.

Signed-off-by: Jason Liu <liujason@google.com>
Google-Bug-Id: 64517700
Change-Id: I0ab0bb33b48352655555fbf8dff5dce6d49d9ef6