ltm: stderr/out to log files, delete unnecessary logs
The log files of the LTM are copied into the results file when it is
being created and uploaded. It's unnecessary to keep the copy of the log
file around locally since it will already be in the GS bucket, so
instead we can just move the file.

Of course, after the file has been moved, the logs for the remaining
finish/cleanup function still need to be sent to where the old file was,
so the log handlers need to be reset.

Additionally, stderr and stdout in the shard/testrunmanager processes
were being dumped into the lighttpd server error logs (mainly an issue
in debugging an uncaught exception). This change moves stdout/stderr
(and by extension, those errors) to the same file as the logfile.

Signed-off-by: Jason Liu <liujason@google.com>
Change-Id: I99342d01b9901c0a8d46d19f94cb6e876fd86cc2