gce-xfstests: check for .config file in GS bucket before running
The .config file in the bucket may not be present when the user runs a
gce-xfstests command. If it isn't, certain options which aren't
propagated through the metadata will not be available to the test
appliance/LTM.

This adds a check before launching to ensure that a config file exists.

Signed-off-by: Jason Liu <liujason@google.com>
Change-Id: Ic1f0a89e47fc75afca444dbe00960324b0a8234f