ltm: serial output still available if shard encounters gs error
The serial output file that the shard creates was renamed to a
different file (unpacked_results_serial). The testrunmanager only
checked whether that file existed to see whether the shard finished
properly or not (if it finished properly, the shard will have deleted
the serial dump instead of renaming it).

Renaming the file is unnecessary, and it's enough to simply delete the
file if the shard successfully finishes.

If the shard exits unexpectedly due to a GS error, the results file
cannot be deleted from GS, and the results for this shard will also be
unavailable in the aggregate results. The serial dump would also be
unavailable, as the shard will not have renamed the serial dump file to
the name that the testrunmanager can see. This change ensures that even
in this event, if there was a serial dump present in a local file for
the shard, it will still be available to the testrunmanager.

Signed-off-by: Jason Liu <liujason@google.com>
Change-Id: Iedaa040e77efc406412b8b7e2a1315103671618b