gce-xfstests ltm: use g1-small and not n1-standard
LTM cpu usage is in small bursts, using a g1-small does not increase
the latency. The memory however is a bit of a concern.

In particular, the memory consumption on the LTM is approximately
50MB flat per test run, with an initial spike of around 50MB per shard,
and a stable consumption of around 20MB per shard.

Using a g1-small (1.7GB of memory) likely supports at most 2
simultaneous ext4/all test runs.

Signed-off-by: Jason Liu <liujason@google.com>
Change-Id: Ief770c86f6b081ec494e3732698f527912933cf2