ltm: add --no-email option to suppress email
It may be wanted for a user to suppress sending an email, or for the
LTM to suppress individual shards sending emails. It is possible to
suppress email by passing an empty string as an argument to the
--email option. This is, however, unwieldy and hard to read if the
command line is printed back out.

This change adds an explicit '--no-email' option to suppress email
sending. It also modifies the LTM to use this option instead of passing
an empty string to --email.

Signed-off-by: Jason Liu <liujason@google.com>
Google-Bug-Id: 64974791
Change-Id: I4d080b26fc0c2f94d5a12248e94b2d16959aa051