gce-xfstests: add multi-zone support
Add support for the various gce-xfstests commands (ls, abort, ssh,
console, etc.) which operate on VM's to work when they are not running
in the default GCE_ZONE for the project.  Since looking up the zone
takes an extra three-quarters of a second, we only do the zone lookup
for VM names which start with "xfstests-ltm-".

Signed-off-by: Theodore Ts'o <tytso@mit.edu>