test-appliance: update to use Debian Stretch
Debian Stretch was released as the new Debian Stable on June 17th,
2017.  The latest version of Stretch, 9.1, was released on July 22,
2017.

It is still possible to build test appliances using Debian Jessie by
passing the command-line arguments "--suite jessie" to the gen-image
script, and the "--jessie" option to the gce-create-image command.

Support for Debian Stretch's backports is not there yet, but as of
this writing there are no packages that we would want or need in
stretch-backports.  (It is likely the packages that we would want or
need from jessie-backports and stretch-backports will be different.
So it likely will be simpler to drop support for Debian Jessie by the
time we want to start using stretch-backports.)

Signed-off-by: Theodore Ts'o <tytso@mit.edu>