gce-xfstests: add new sub-command: copy-image
Creating a temporary disk to copy an image from one GCE project to
another is more efficent than using gce-export-image and
gce-import-image which requires starting an VM and copying the bits to
a bucket in Google Cloud Sturage.

The copied image also will be more accurate since the description,
labels, licenses, and guestOSFeatures will be preserved.

The two primary uses of this command will be (a) by gce-xfstests
developers who making releases of the test appliance image to the
xfstests-cloud project, and (b) users who need their own copy of the
test appliance image so the LTM server (which is running using a
service account) can have access to the image.

Signed-off-by: Theodore Ts'o <tytso@mit.edu>