test-appliance: git rid of sendgrid username/password support
It's simpler and more secure to use a Sendgrid API key, and that's the
only thing which is documented at this point.  Removing it also means
we don't need to include postfix in the test appliance VM.

Also export GCE_REPORT_EMAIL in the gce_xfstests.config file stored in
Google Cloud Storage bucket so that if "gce-xfstests ltm" is run on
the LTM server, the resulting test report will be emailed to the user.

Signed-off-by: Theodore Ts'o <tytso@mit.edu>