test-appliance: improve image generation for chroot tar files
With Debian Stretch, systemctl has a DT_RPATH for its systemd shared
library, and fakechroot doesn't seem to handle this correctly.  So
this causes debootstrap to fake when running gen-image --out-tar using
fakechroot.  So teach gen-image so it can work without using
fakechroot if it is running with root privileges.

Also fix gen-image so that it can work correctly when using Qemu user
emulation (e.g., when we are building an armhf binary image on an x86
host platform).

Signed-off-by: Theodore Ts'o <tytso@mit.edu>