setup-buildchroot: fix copying passwd and group entries
create_chroot() is now supposed to add $CHROOT_USER to /etc/passwd in
the chroot.  However, due to a bad grep pattern, it would add multiple
entries if $CHROOT_USER happened to be a suffix of another username.
Also the passwd record is not guaranteed to actually be in the file
/etc/passwd.  Fix it by using 'getent passwd $CHROOT_USER' instead.

A similar problem existed for the group entry, but also the
$CHROOT_USER's primary group is not guaranteed to have the same name as
the user.  Fix that by using 'id -g' to look up the primary group.

Signed-off-by: Eric Biggers <ebiggers@google.com>
Signed-off-by: Theodore Ts'o <tytso@mit.edu>