test-appliance: fix use of backports
The backport-packages stuff was for jessie, but falls over now that
we're using stretch.  So instead make this debian flavor dependent, so
that if we need backports in the future for stretch we'll have
backport-packages-stretch, and likewise if people want to build a
jessie image everything will still work as it did before.

Signed-off-by: Josef Bacik <josef@toxicpanda.com>
Signed-off-by: Theodore Ts'o <tytso@mit.edu>