kvm-xfstests: add result archive generation
It would be nice have an ability to fetch detailed logs from result vdisk image
but in order to do that we need to mount it, which annoying. Even more
this not works for environments where we have not mount permissions or
access to /dev/loop.  Let's archive tests results to single file and fetch
it from disk image via debugfs.
Example:
#./kvm-xfstests.sh --archive -c 4k generic/001
...
result  in ./logs/resutls-201704081703.tar.xz
logfile in ./logs/log.201704081703

Signed-off-by: Dmitry Monakhov <dmonakhov@openvz.org>