docker: parameterize the URLs used for the root_fs and sample kernel
If there is not a root_fs.img in the test-appliance directory, we
download one from www.kernel.org.  Make this be parameterized so it's
easier to read and so it can be overridden using docker build's
--build-arg option.

Similarly, parameterize the location of the Fedora kernel that we use
as a sample kernel so it can be overriden using docker run's --env
option.

Signed-off-by: Theodore Ts'o <tytso@mit.edu>