test-appliance ltm: move the base64 decoding up to app.py
Previously the LTM python scripts were passing the base-64 encoded
strings around, to the point where the Sharder class was base-64
encoding a string only to have the Shard class base-64 decode it.

Base-64 strings are larger, more unwieldy, harder to deal with when
debugging, and only have one benefit -- they won't get confuse parsers
that split on whitespace.  So change things we do the base-64 decode
at top-level of the webapp, and then just pass the raw string around.

Signed-off-by: Theodore Ts'o <tytso@mit.edu>