test-appliance: exclude generic/382 from the ext3 configuration
The generic/382 test fails on ext3 configuration because the quota
report, which includs the number of blocks used after writing a 30 meg
file, is in the golden output.  Unfortunately this doesn't take into
account the indirect blocks which are needed when we are in ext3 mode
and the extents feature is disabled.

It's not clear whether this should be considered a test bug or not,
but at least for now, suppress this test in the ext3 configuration for
{kvm,gce,android}-xfstests.

Signed-off-by: Theodore Ts'o <tytso@mit.edu>