test-appliance: use --auto-deconfigure when installing extra packages
Trying to build the KVM or GCE test appliances with Debian packages
built from the latest e2fsprogs git tree placed into the
test-appliance/debs/ directory fails because of the recent renamings
from libcomerr2 to libcom-err2 and e2fslibs to libext2fs2, e.g.:

    dpkg: regarding .../libcom-err2_1.44.0~rc2-1_amd64.deb containing libcom-err2:amd64:
     libcom-err2 breaks libcomerr2 (<< 1.43.9-1~)
      libcomerr2:amd64 (version 1.43.4-2) is present and installed.

    dpkg: error processing archive /run/libcom-err2_1.44.0~rc2-1_amd64.deb (--install):
     installing libcom-err2:amd64 would break libcomerr2:amd64, and
     deconfiguration is not permitted (--auto-deconfigure might help)

Add the --auto-deconfigure argument to make it work.  I'm not sure it's
the best solution, but it works.

Signed-off-by: Eric Biggers <ebiggers@google.com>
Signed-off-by: Theodore Ts'o <tytso@mit.edu>