Accelerate builds on qemu user chroots
Certain binaries (most notably the shell, coreutils, and sed) can be
built for the native architecture and put in the /usr/local/bin-accel
directory.  The build process will use them automatically where it is
safe to do so.  (For example, it is not safe to use them if fakechroot
or fakeroot is being used.)

Signed-off-by: Theodore Ts'o <tytso@mit.edu>