kvm-xfstests: use the host cpu type for the guest by default
Previously we were not passing the -cpu option to qemu, so it
defaulted to the "QEMU Virtual CPU" which is the lowest common
denominator.  Use the "host" cpu type by default, which allows the
guest to use the CPU features of the host CPU.

Also add the --cpu-type option to kvm-xfstests to override the cpu
type.

The default CPU type can also be set by defining $CPU_TYPE in
~/.config/kvm-xfstests.

Signed-off-by: Theodore Ts'o <tytso@mit.edu>