test-appliance, kvm-xfstests: fix up networking support
Networking support had bitrotted somewhat since it hadn't been used in
a while.  Fix it up so it works with newer versions of systemd which
would try to rename eth0.  Also fix up NAT support when kvm-xfstests
is run as root.  Also fix the network config so that if the image is
built with networking support enabled, and kvm-xfstests is run without
networking enabled, that the VM won't cause the boot sequence to fail
due to the missing networking interface.

Signed-off-by: Theodore Ts'o <tytso@mit.edu>