{kvm,gce}-xfstests: use the kernel in the current directory by default
If the --kernel parmaeter is not explicitly specified, then first try
to see if the kbuild configuration file is present, use the specified
build directory to find the kernel.  Otherwise, if the current
directory is the top-level of a kernel build tree where a kernel can
be found, use it.  If neither these is true, at that point go back to
the previous behavior of using the compiled-in default.

Signed-off-by: Theodore Ts'o <tytso@mit.edu>