test-appliance: change variable set by get_one_fs_config from $TC to $i
As part of some code refactoring, the bash shell function
get_one_fs_config was created, and calling makes changes to a number
of shell variables; specifically, it removes a test config from
FSTESTCFG, and as output sets the variables FS and i.  Side effecting
a global variable is a bit icky, but it is idiomatic in bash and tcl
scripts.  If we *are* going to do this, though, we should use
something less likely to be used locally, so change get_one_fs_config
to set TC to be the test config for the file system type FS which is
to be tested.

Signed-off-by: Theodore Ts'o <tytso@mit.edu>