test-appliance: save gce-setup-scratch and gce-setup-nfs to /results
These two scripts do some complex things that may fail due to issues
with the gcloud command failing due to GCE issues or the gcloud
command changing in some way.  So in order to debug what might be
transient failures, it's important to save them to a place where they
can be examined if a test fails mysteriously.

Signed-off-by: Theodore Ts'o <tytso@mit.edu>