test-appliance: add the nfs/filestore file system configuration
This commit cleans up how we support NFS file systems.  The
nfs/filestore is the officially supported way to test the NFS client
versus a Google Compute Engine's NFS Filer, Filestore.  It's still
possible to specify the nfs server of fileserver and use nfs/client,
but this may not work in the future; the primary use case for
nfs/client is testing against generic NFS servers.

Signed-off-by: Theodore Ts'o <tytso@mit.edu>