test-appliance: use the nosharecache mount option for nfs/filestore
We are using the same NFS export for the test and scratch mounts, and
since GCE Filestore doesn't support setting separate fsid for separate
directories on the same filers, fix generic/306 and generic/307
failures by using nosharecache.  This allows a r/o mount for one mount
not to affect the r/o status of the other.

Signed-off-by: Theodore Ts'o <tytso@mit.edu>