test-appliance: update Python packages to the lastest versions
Update Python packages to the lastest versions to fix various security
issues --- except for flup, where flup 1.0.3 breaks things.  There are
some reports that it is problematic with Python 2, but given that flup
doesn't seem to have great documentation; doesn't seem to be updated
regularly or consistently; and doesn't have a place where bugs can be
reported, we should consider replacing flup with something else.  One
possibility is to replace lighttpd+flup with a hosted WSGI container
such as Gunicorn (which is Python based and which can support https).

Signed-off-by: Theodore Ts'o <tytso@mit.edu>