image-utils: improve label generation algorithm
Fix the label generation code so that something which looks like a git
commit (where the first three characters are hex digits) are prefixed
with the letter 'g'.  If the first letter is a number, then prefix it
with the letter 'v', since GCE labels must begin with a lower-case
alpha character.  Otherwas use the version in the git-versions file.

The old algorithm would transform git versions which consisted only of
a the git tag, such as "v4.19.0" to "gv4.19.0", which was ugly,
confusing and unnecessary.

Signed-off-by: Theodore Ts'o <tytso@mit.edu>