test-appliance: wait for set-disk-auto-delete to complete
In gce-setup-scratch, we need to wait for the command to set the
scratch disk for auto-deletion.  Otherwise the gce-setup-scratch
script will often finish first, and this will cause the
set-disk-auto-delete commant to get killed prematurely.

Also save the gce-setup-* log files in /var/log instead of /results,
since most of the time we don't need to keep them in the archived
results tarfiles.

Signed-off-by: Theodore Ts'o <tytso@mit.edu>