kernel-configs: add misc. file systems and enable UEFI booting
Booting the x86_64 kernels on Google Compute Engine with Debian Buster
requires that the kernel be built with UEFI support enabled.  So
enable it for 3.10 and newer kernels.

Also enable JFS, reisefs, and NFS, which the test appliance can now
test.  Also enabled is nilfs, although for now the xfstests suite
doesn't support nilfs.

Signed-off-by: Theodore Ts'o <tytso@mit.edu>