test-appliance: exclude generic/471
The test generic/471 tests the RWF_NOWAIT flag; however
how it is supposed to work with file systems is disputed,
and not all device drivers support it.  As a result
it doesn't work if an LVM device driver is in use (as is the
case with gce-xfstests).  So let's suppress it for now.  For
more details see:

https://lore.kernel.org/linux-block/20190723220502.GX7777@dread.disaster.area/

Signed-off-by: Theodore Ts'o <tytso@mit.edu>