test-appliance: teach gce-kexec so it can be re-run after a reboot
A lot of the setup work done by gce-kexec doesn't need to be redone if
the test appliance is rebooted.  (For example, if we are using a
preemptible VM.)  In particular, we want to use the kernel that was
initially copied to /root/bzImage, since version in GCS may have been
later removed or overwritten.

Signed-off-by: Theodore Ts'o <tytso@mit.edu>