test-appliance: disable man-db's auto-update
At the end of building the test appliance, we delete all of the files
in /var/cache/man.  So if a developer needs to install some package
which comes with a man page, man-db will rebuild /var/cache/man from
scratch, and that can take quite a long time.  So disable man-db's
auto-update.

Signed-off-by: Theodore Ts'o <tytso@mit.edu>