test-appliance: fetch the gce_xfstests.config file in gce-setup
If the file is not present, fetch it so we can be sure it's there.
The comment that we would have fetched it from a previous boot is not
necessarily true, so remove it.

Also unconditionally fetch the file in gce-fetch-gs-files (used by the LTM
setup path), since if the file isn't present, skipping its download
isn't going to help and don't bother checking to see if the
gce_xfstests.config file exists before we try to copy it.

Finally, we need to export SENDGRID_API_KEY since gce_xfstests.config
doesn't do this for us, when previously /usr/local/lib/sg_key did.

Fixes: 052529c37714 ("test-appliance: get rid of /usr/local/lib/sg_key")
Signed-off-by: Theodore Ts'o <tytso@mit.edu>