build-all: allow builds to be more repeatable regardless of build environment
Disable use of CPU-specific optimizations based on the build platform
when building fio; they may not exist where we plan to run the test
appliance.

Forcibly disable selinux and systemd system when building util-linux;
we don't need it, and that way we know we will be consistent when
building outside of a build chroot.

Allow the parent process of build-all to set some CPU-specific cflags,
such as forcibly building with a subset of CPU optimizations.

Signed-off-by: Theodore Ts'o <tytso@mit.edu>