test-appliance: disable ext4 tests that are incompatible with test scenarios
ext4/044 tries to mount a file system explicitly as ext3, so file
system configs which use dioread_nolock, nodioread_nolock, or
nodelalloc won't work.  In addition ext3 doesn't support file system
encryption.

ext4/045 creates a large number of directories, so using a 1k block
won't work because the max htree depth limitation (unless the
large_dir feature is enabled).

generic/381 uses quotas, so it doesn't work with the
test_dummy_encryption mount option enabled, as is the case in the
encrypt.

generic/556 is a casefolding test, which for now doesn't work
encryption enabled.

Signed-off-by: Theodore Ts'o <tytso@mit.edu>