test-appliance: update VM pricing table in gce-xfstests
Add the e2 machine types since they are cheaper than the n1 machine
types with roughly the same performance.  The n2 machine types are
more expensive per CPU hour, and is faster, and is a slight net win
over the n1 machine types --- but the e2 VM's are better.  Comparing
the runtime and costs of:

	gce-xfstests -c ext4/4k -g auto --machtype <machtype>

GCE MachType        Time        Cost         Savings
n1-standard-2       6178s     20.5 cents        ---
e2-standard-2       5942s     15.0 cents       26.8%
n2-standard-2       5513s     18.6 cents        9.3%

This patch doesn't actually enable the automatic selection of the e2
VM's, since we need to be careful; not all of the GCE zones support
the e2 machine types.

(All of the prices are for the us-central1 zone but it's really the
relative differences in prices which are useful for selecitng the
appropriate machine type.)

Signed-off-by: Theodore Ts'o <tytso@mit.edu>