test-appliance: use a dynamic algorithm to pick the GCE machine type
Different GCE zones support different GCE machine types.  So we need
to fetch the GCE region information and use that to decide the best
machine type to use.  In general this will be the e2-standard- VM's,
but there may be places where only the old n1-standard VM's which are
the best.  The n2 and n2d machine types are not cost effective for
gce-xfstests so they are not considered.

Signed-off-by: Theodore Ts'o <tytso@mit.edu>