use internal ip for ssh and curl if running within the same gce project
This patch tries to get around some gce firewall rules that block ssh or
http access from outside the gce project. The script checks whether it's
running on a GCE VM, then use the GCE metadata or internal ip of this VM
to determine whether we communicate with new VMs with the internal ip.

Signed-off-by: Tim Shen <xyshen@google.com>
Signed-off-by: Theodore Ts'o <tytso@mit.edu>