gen-tarball: add 'set -e'
'make install' in xfstests was temporarily broken (it was broken by
commit 10f3f9707cca and fixed by commit 1de2b58b8e29).  gen-tarball
didn't notice this, resulting in test appliances being "successfully"
built but not actually containing all the xfstests files.

Fix gen-tarball to check for errors by adding 'set -e'.  To make this
work, adjust a few commands that are expected to fail.

Signed-off-by: Eric Biggers <ebiggers@google.com>
Signed-off-by: Theodore Ts'o <tytso@mit.edu>