test-appliance: add support to use a non-default account via $GCE_ACCOUNT
This allows specifying a hard-coded account to be used when running
gcloud, so gce-xfstests does not have to depend on a specific gcloud
configuration.

As part of this, refactor all calls to gcloud to use the bash
functions run_gcloud, run_gcloud_always, and run_gcloud_prj.

Also support enabling --no-action as the first argument so that we
debug the gce-xfstests commands which do not run a test.

Signed-off-by: Theodore Ts'o <tytso@mit.edu>