gce-create-image: optimize test appliance building
Reduce the amount of time to build the appliance by approximately 30%
(from 10 to 7 minutes) and reduces the time running the xfstests-bld
VM from 550 to 350 seconds.

This is done by running "go get" locally and then uploading the
gigabyte of go package sources to GCS as a cache.  To speed things up,
we will pigz, a parallel gzip compressor.

Signed-off-by: Theodore Ts'o <tytso@mit.edu>