test-appliance: set the boot disk to be auto-delete after finishing setup
There is a kernel bug which can cause the VM to crash hang after the
kexec.  In that case, the LTM server will kill the VM, and in that
case there's point keeping the boot disk chewing up PD storage
charges.  So launch the VM with the boot disk set to auto-delete, and
clear the the auto-delete flag at the end of /usr/local/lib/gce-setup.

Signed-off-by: Theodore Ts'o <tytso@mit.edu>