test-appliance: unify the .deb/bzImage kernel load path in the GCE Test VM
The way the test parametsr were passed when loading a kernel.deb was
different from what was used by kvm-xfstess and gce-xfstests with
kexec.  Unfortunately the kernel.deb load path didn't support passing
kernel options to the kernel under test --- which is needed when
testing DAX, and when the developer uses the gce-xfstests options -n,
-r, or -o.

There was also a bug in the kernel.deb path where fstestapi was not
correctly getting split, which caused warnings when the test VM was
attempting to validate the test API version.

We now set the kernel boot command line options via
GRUB_CMDLINE_LINUX, so there is a consistent way of plumbing test
parameters to the kernel under test.

Fixes: fae31d5cafa3 ("test-appliance: add support to boot kernel via deb")
Signed-off-by: Theodore Ts'o <tytso@mit.edu>