test-appliance: include the built man pages in the GCE test appliance
The gce-xfstests test appliance includes the man utility, so in order
to allow the xfs tests which check to make sure the man pages for
xfsprogs document all of the supported command to succeed, we need to
include the manual pages in xfstests/share/man.

They are still not included in the KVM test appliance since it doesn't
include the man utility and we are trying to keep the KVM test
appliance as small as possible.

Signed-off-by: Theodore Ts'o <tytso@mit.edu>