gce-xfstests: add support to specify the GCE PD specification
This adds the gce-xfstests option --gce-disk-spec which allows the
user to request the test VM to create the PD scratch disk of a
specific type and size.   The syntax is:

   [<scope>/][type][/<size>]

Where scope can be either "zonal" or "regional", type can be
"pd-standard", "pd-ssd", "pd-balanced", or "pd-extreme", and size of
the disk in gigabytes.  If the scope or size is not specified, we
default to "zonal" and "100GB", respectively.  If the gce-disk-spec is
not specified the default type is pd-ssd.

Signed-off-by: Theodore Ts'o <tytso@mit.edu>