gce_xfstests: add dashboard support
Add support to visualize the test results. This visualization is
implemented using google cloud run[1]. The dashboard is implemented in
a containerized python script. gce-xfstests dashboard launch command
builds and uploads the container on cloud run, which is google cloud's
low cost serverless platform.

This patch adds following new commands to gce-xfstests:

gce-xfstests launch-dashboard - Launch the dashboard as an app on
cloud run

Optionally, --local flag can be passed to the above command which
would launch the gce-xfstests dashboard locally.

[1] https://cloud.google.com/run

[ Updated to use new service account setup framework and fix missing
  the role permissions granted to the service accouint for the
  dashboard.  --TYT ]

Signed-off-by: Harshad Shirwadkar <harshadshirwadkar@gmail.com>
Signed-off-by: Theodore Ts'o <tytso@mit.edu>