test-appliance: add fast_commit to the ext4/adv test configuration
The fast_commit file system feature is a compat feature, so it's safe
to enable it in the adv test configuration --- it will be ignored on
kernels that don't support the feature.

Also remove from the adv feature those features which are enabled by
default for all ext4 file systems, since they don't actually do
anything with current versions of e2fsprogs.

Signed-off-by: Theodore Ts'o <tytso@mit.edu>