test-appliance: always create all of the partitions for gce-xfstests
Previously compute_partition_sizes tried to be smart by figuring out
which parititions would actually be needed by the fs configs which
would be run.  The original reason for doing this was to save money by
creating a PD SSD that was only as large enough needed for the test.
However, it turned out that we always needed to create PD of a minimum
size so there would be enough IOPS provisioned to make the tests run
in a reasonable time (otherwise, the increased run time would result
in the CPU time needed for the VM swamping the money saved by using a
smaller PD disk).

Reduce the complexity needed in keeping compute_partition_sizes in
sync with runtests.sh by simply assuming that we will need all of the
possible partitions for running tests.

Signed-off-by: Theodore Ts'o <tytso@mit.edu>