test-appliance: add ext4/044 to adv.exclude
Test ext4/044 fails when run in the adv test case because it explicitly
attempts to mount a test file system created with the inline_data
feature as ext3.  The inline_data feature and ext3 are incompatible,
and the mount attempt fails.

This test did not fail in earlier xfstests-bld versions because the
features included in adv were different.  In particular, the 64bit
feature was applied, and this had an unfortunate side effect.
Because the 64bit feature requires extents, and because the test
attempts to create an ext3 file system, the initial attempt actually
failed.  This was hidden by behavior in the xfstest function
_scratch_do_mkfs, which then attempted to create the file system without
the supplied "extra" mkfs options (those supplied for the adv test case).
So, the test file system was not created with inline_data, the explicit
attempt to mount the test file system as ext3 succeeded, and the test
passed without testing anything particular to the adv test case.

Signed-off-by: Eric Whitney <enwlinux@gmail.com>
Signed-off-by: Theodore Ts'o <tytso@mit.edu>