build-all, setup-buildchroot: build xfsprogs with unicode support enabled
In order to build xfsprogs with unicode support, it turns out that we
need to link xfsprogs with system libraries using dynamic linking.
The libicu library seems to build such that its C++ symbols don't link
correctly when using static linking.

Upon investigations, build-all was building some packages completely
statically linked, while other packages were linked dynamically.  The
reason for using static linking is so that the binaries in the
xfstests.tar.gz aren't dependent on the shared libraries available in
the run-time environment where they are unpacked.  Unfortunately, when
some of the binaries are dynamically linked, and some are statically
linked, we get the worst of both worlds.  The statically linked
binaries take more space, and dynamically linked binaries introduce a
dependency on the system shared libraries.

So build-all now builds all of the binaries dynamically linked with
the system libraries (although the binaries still statically link with
the binaries built by xfstests-bld, so that we don't need to worry
about rpath and the search path for shared libraries).  This reduces
the size of the binaries and tests in xfstests directory from 64
megabytes to 53 megabytes.

The build-all script has a --static option to build nearly[1]
everything completely statically linked.  The resulting size of the
xfstests directory is 187 megabytes.

[1] Unfortuantely, the fsverity and the ima-evm-utils package can't
currently be built completely statically due to shortcomings in their
build scripts.  Hopefully these limitations can be lifted in the
future, in particular for fsverity, since the fix is simple and
fsverity is required external dependency, whereas ima-evm-utils is
optional.

Signed-off-by: Theodore Ts'o <tytso@mit.edu>