gce-xfstests: add commands to cache or regenerate the machtype file
Generating the machtype file can be quite expense, especially if the
LTM server is spawning a dozen or more gce-xfstests commands, which
race to generate the machtype file.  This can significantly slows down
the time it takes the first time LTM spawns a test if it needs to
spawn a large number of VM's.

To address this when starting up the LTM server, run "gce-xfstests
cache-machtype-file" so that the cached file is prepopulated before
the LTM server attempts to launch a test run.

Signed-off-by: Theodore Ts'o <tytso@mit.edu>