test-appliance: remove convenience fstab entries
Fstab entries for the xfstests filesystems interfere with xfstests
because they change the behavior of 'mount -o remount MOUNTPOINT' to
remount relative to the mount options from the fstab rather than the
actual mount options.  For example:

With fstab:
	$ mount -o lazytime /dev/vdc /vdc
	$ mount -o remount /vdc
	$ findmnt /vdc
	TARGET SOURCE   FSTYPE OPTIONS
	/vdc   /dev/vdc ext4   rw,relatime

Without fstab:
	$ rm /etc/fstab
	$ mount -o lazytime /dev/vdc /vdc
	$ mount -o remount /vdc
	$ findmnt /vdc
	TARGET SOURCE   FSTYPE OPTIONS
	/vdc   /dev/vdc ext4   rw,relatime,lazytime

So the lazytime option was lost on remount due to the fstab.  This
happens with other mount options too; lazytime is just an example.

This breaks xfstest ext4/053.

Remove the fstab entries, as they were only needed to make it slightly
easier to manually mount filesystems in an interactive shell.  So 'mount
/vdc' will no longer work, but 'mount /dev/vdc /vdc' will still work.

Alternatively we could make xfstests always use
'mount --options-source=mtab' instead of just 'mount', but xfstests
calls 'mount' in a bunch of different places.  We could still do it
anyway, though; let me know if that would be preferred.

Signed-off-by: Eric Biggers <ebiggers@google.com>
Signed-off-by: Theodore Ts'o <tytso@mit.edu>